requirements.txt
py2neo
neo4j-connector
Cython
kivy
wmd
spacy
scispacy
https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.2.0/en_core_sci_md-0.2.0.tar.gz
pipetools
regex
more_itertools
addict
numpy
numpy_indexed
suffix_trees
scipy
cytoolz
hdbscan
pyprover
pyxdameraulevenshtein
ray
biopython
tabulate
fuzzywuzzy[speedup]
git+https://github.com/c0ntradicti0n/Distinctiopus4.git
